Align delete code with existing UI extensions lib
